# Conflicts:
#	backend/src/ServiceReleaseManager.Api/Endpoints/Organisations/ListServiceTemplates.cs
#	backend/src/ServiceReleaseManager.Api/Endpoints/Services/Create.cs

+++ backend/src/ServiceReleaseManager.Api/Endpoints/Locales/Delete.cs
@@ -1,5 +1,7 @@
 ﻿using Ardalis.Result.AspNetCore;
 using Microsoft.AspNetCore.Mvc;
+using ServiceReleaseManager.Api.Authorization;
+using ServiceReleaseManager.Api.Authorization.Operations.Locale;
 using ServiceReleaseManager.Core.Interfaces;
 using Swashbuckle.AspNetCore.Annotations;
 
@@ -8,10 +10,13 @@ namespace ServiceReleaseManager.Api.Endpoints.Locales;
 public class Delete : EndpointBase.WithRequest<DeleteLocaleRequest>.WithoutResult
 {
   private readonly ILocaleService _localeService;
+  private readonly IServiceManagerAuthorizationService _authorizationService;
 
-  public Delete(ILocaleService localeService)
+  public Delete(ILocaleService localeService,
+    IServiceManagerAuthorizationService authorizationService)
   {
     _localeService = localeService;
+    _authorizationService = authorizationService;
   }
 
   [HttpDelete(DeleteLocaleRequest.Route)]
@@ -27,6 +32,15 @@ public class Delete : EndpointBase.WithRequest<DeleteLocaleRequest>.WithoutResul
     [FromRoute] DeleteLocaleRequest request,
     CancellationToken cancellationToken = new())
   {
+    var locale = await _localeService.GetById(request.LocaleId, cancellationToken);
+
+    if (!locale.IsSuccess || !await _authorizationService.EvaluateServiceAuthorization(User,
+          locale.Value.ServiceId,
+          LocaleOperations.Locale_Delete, cancellationToken))
+    {
+      return NotFound();
+    }
+
     var result = await _localeService.Delete(request.LocaleId, cancellationToken);
 
     return this.ToActionResult(result);
